+++ extensions/fantastical/src/isInstalled.ts
@@ -1,0 +1,5 @@
+import { getApplications } from "@raycast/api";
+
+export async function isInstalled() {
+  return (await getApplications()).find((app) => app.name === "Fantastical") != undefined ? true : false;
+}