+++ extensions/obsidian/src/components/NoteListObsidian.tsx
@@ -2,34 +2,34 @@ import { showToast, Toast, getPreferenceValues } from "@raycast/api";
 import React, { useEffect, useMemo, useState } from "react";
 import fs from "fs";
 
-import NoteLoader from "../utils/NoteLoader";
-import { Note, Vault, SearchNotePreferences } from "../utils/interfaces";
+import { useNotes } from "../utils/cache";
+import { Note, Vault, SearchNotePreferences, SearchArguments } from "../utils/interfaces";
 import { NoteList } from "./NoteList";
-import { filterNotes, getListOfTags } from "../utils/utils";
+import { filterNotes } from "../utils/search";
 import { MAX_RENDERED_NOTES, NoteAction } from "../utils/constants";
 import { NoteActions, OpenNoteActions } from "../utils/actions";
+import { tagsForNotes } from "../utils/yaml";
 
-export function NoteListObsidian(props: { vault: Vault; showTitle: boolean }) {
+export function NoteListObsidian(props: { vault: Vault; showTitle: boolean; searchArguments: SearchArguments }) {
   const { searchContent } = getPreferenceValues<SearchNotePreferences>();
 
-  const { showTitle, vault } = props;
+  const { showTitle, vault, searchArguments } = props;
   const [notes, setNotes] = useState<Note[]>([]);
   const [allNotes, setAllNotes] = useState<Note[]>([]);
 
-  const [input, setInput] = useState<string>("");
+  const [input, setInput] = useState<string>(searchArguments.searchArgument || "");
   const list = useMemo(() => filterNotes(notes, input, searchContent), [notes, input]);
 
   function onDelete(note: Note) {
     setNotes(notes.filter((n) => n.path !== note.path));
+    setAllNotes(allNotes.filter((n) => n.path !== note.path));
   }
 
   useEffect(() => {
     async function fetch() {
       try {
         await fs.promises.access(vault.path + "/.");
-        const nl = new NoteLoader(vault);
-        const _notes = nl.loadNotes();
-
+        const _notes = useNotes(vault);
         setNotes(_notes);
         setAllNotes(_notes);
       } catch (error) {
@@ -43,7 +43,7 @@ export function NoteListObsidian(props: { vault: Vault; showTitle: boolean }) {
     fetch();
   }, []);
 
-  const tags = getListOfTags(allNotes);
+  const tags = tagsForNotes(allNotes);
 
   return (
     <NoteList
@@ -55,6 +55,7 @@ export function NoteListObsidian(props: { vault: Vault; showTitle: boolean }) {
       onSearchChange={setInput}
       onDelete={onDelete}
       tags={tags}
+      searchArguments={searchArguments}
       action={(note: Note, vault: Vault, actionCallback: (action: NoteAction) => void) => {
         return (
           <React.Fragment>