+++ extensions/pipe-commands/assets/commands/change-case/lowercase.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 
+# @raycast.schemaVersion 1
 # @raycast.title Switch to Lowercase
 # @raycast.packageName Text Actions
-# @raycast.mode replace
-# @raycast.argument1 {"type": "text"}
+# @raycast.mode pipe
 
 import sys
 
-sys.stdout.write(sys.argv[1].lower())
+sys.stdout.write(sys.stdin.read().lower())