+++ extensions/pipe-commands/assets/commands/text/count-characters.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+# @raycast.title Count Characters
+# @raycast.mode silent
+# @raycast.icon List
+# @raycast.argument1 {"type": "text", "percentEncoded": false}
+
+echo -n "${#1}"