+++ extensions/spotify-player/src/searchPlaylists.tsx
@@ -1,9 +1,10 @@
-import { Action, ActionPanel, Icon, Image, List, showToast, Toast } from "@raycast/api";
+import { List, showToast, Toast } from "@raycast/api";
 import { useEffect, useState } from "react";
-import { play, playShuffled, usePlaylistSearch } from "./client/client";
-import { isSpotifyInstalled } from "./client/utils";
+import { usePlaylistSearch } from "./spotify/client";
+import { isSpotifyInstalled } from "./utils";
+import PlaylistItem from "./components/PlaylistListItem";
 
-export default function SpotifyList() {
+export default function SearchPlaylists() {
   const [searchText, setSearchText] = useState<string>();
   const [spotifyInstalled, setSpotifyInstalled] = useState<boolean>(false);
   const response = usePlaylistSearch(searchText);
@@ -33,60 +34,5 @@ export default function SpotifyList() {
         <PlaylistItem key={p.id} playlist={p} spotifyInstalled={spotifyInstalled} />
       ))}
     </List>
-  );
-}
-
-function PlaylistItem(props: { playlist: SpotifyApi.PlaylistObjectSimplified; spotifyInstalled: boolean }) {
-  const playlist = props.playlist;
-
-  const spotifyInstalled = props.spotifyInstalled;
-  const imageURL = playlist.images[playlist.images.length - 1]?.url;
-  const icon: Image.ImageLike = {
-    source: imageURL ?? Icon.TextDocument,
-    mask: Image.Mask.Circle,
-  };
-
-  const title = playlist.name;
-  const subtitle = playlist.owner.display_name;
-  return (
-    <List.Item
-      title={title}
-      subtitle={subtitle}
-      accessories={[{ text: `${playlist.tracks.total.toString()} songs`, tooltip: "number of tracks" }]}
-      icon={icon}
-      actions={
-        <ActionPanel title={title}>
-          <Action
-            title="Play"
-            icon={Icon.Play}
-            onAction={() => {
-              play(undefined, playlist.uri);
-            }}
-          />
-          <Action
-            icon={Icon.Shuffle}
-            title="Play Shuffled"
-            onAction={() => {
-              playShuffled(playlist.uri);
-            }}
-          />
-          <Action.OpenInBrowser
-            title={`Show Playlist (${playlist.name.trim()})`}
-            url={spotifyInstalled ? `spotify:playlist:${playlist.id}` : playlist.external_urls.spotify}
-            icon={icon}
-            shortcut={{ modifiers: ["cmd"], key: "a" }}
-          />
-          <Action.OpenInBrowser
-            title="Show Artist"
-            url={spotifyInstalled ? `spotify:artist:${playlist.owner.id}` : playlist.owner.external_urls.spotify}
-          />
-          <Action.CopyToClipboard
-            title="Copy URL"
-            content={playlist.external_urls.spotify}
-            shortcut={{ modifiers: ["cmd"], key: "." }}
-          />
-        </ActionPanel>
-      }
-    />
   );
 }