+++ extensions/xcode/src/create-swift-playground.command.tsx
@@ -1,0 +1,5 @@
+import { XcodeCreateSwiftPlaygroundForm } from "./components/xcode-create-swift-playground/xcode-create-swift-playground-form.component";
+
+export default () => {
+  return <XcodeCreateSwiftPlaygroundForm />;
+};