+++ extensions/hackmd/CHANGELOG.md
@@ -1,3 +1,8 @@
 # HackMD Changelog
 
+## [Fixes] - 2022-08-09
+
+- Cached user object to make team selection state preserved in the next command run
+- Fix isLoading state for NoteList
+
 ## [Initial Version] - 2022-08-03